@@ -1,14 +1,15 @@
-﻿using LogisticsSuite.Dispatch.Handlers;
-using LogisticsSuite.Dispatch.Hubs;
-using LogisticsSuite.Dispatch.Services;
+﻿using LogisticsSuite.Backend.Handlers;
+using LogisticsSuite.Backend.Hubs;
+using LogisticsSuite.Backend.Services;
 using LogisticsSuite.Infrastructure.Messages;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
-namespace LogisticsSuite.Dispatch
+namespace LogisticsSuite.Backend
 {
 	public class Startup
 	{
@@ -17,7 +18,7 @@ namespace LogisticsSuite.Dispatch
 		public Startup(IConfiguration configuration) => this.configuration = configuration;
 
 		// This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-		public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+		public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
 		{
 			if (env.IsDevelopment())
 			{
@@ -32,8 +33,8 @@ namespace LogisticsSuite.Dispatch
 				.Register<WebOrderReleasedMessage, WebOrderReleasedMessageHandler>()
 				.Register<CallOrderReleasedMessage, CallOrderReleasedMessageHandler>();
 			app.UseBatchServices();
-			app.UseSignalR(routes => routes.MapHub<MonitorHub>("/ws"));
-			app.UseMvc();
+			app.UseRouting();
+			app.UseEndpoints(endpoints => { endpoints.MapHub<MonitorHub>("/ws"); });
 		}
 
 		// This method gets called by the runtime. Use this method to add services to the container.
@@ -48,7 +49,7 @@ namespace LogisticsSuite.Dispatch
 				.AddTransient<WebOrderReleasedMessageHandler>()
 				.AddTransient<CallOrderReleasedMessageHandler>()
 				.AddBatchService<IMonitoringService, MonitoringService>();
-			services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+			services.AddControllers().SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
 			services.AddSignalR().AddJsonProtocol();
 		}
 	}
