@@ -3,11 +3,8 @@ using LogisticsSuite.Warehouse.Handlers;
 using LogisticsSuite.Warehouse.Repositories;
 using LogisticsSuite.Warehouse.Services;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using IHostingEnvironment = Microsoft.AspNetCore.Hosting.IHostingEnvironment;
 
 namespace LogisticsSuite.Warehouse
 {
@@ -18,19 +15,13 @@ namespace LogisticsSuite.Warehouse
 		public Startup(IConfiguration configuration) => this.configuration = configuration;
 
 		// This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-		public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+		public void Configure(IApplicationBuilder app)
 		{
-			if (env.IsDevelopment())
-			{
-				app.UseDeveloperExceptionPage();
-			}
-
 			app.UseMessaging()
 				.Register<OrderReleasedMessage, OrderReleasedMessageHandler>()
 				.Register<ReplenishedMessage, ReplenishedMessageHandler>()
 				.Register<DelayChangeRequestedMessage, DelayChangeRequestedMessageHandler>();
 			app.UseBatchServices();
-			app.UseMvc();
 		}
 
 		// This method gets called by the runtime. Use this method to add services to the container.
@@ -45,8 +36,6 @@ namespace LogisticsSuite.Warehouse
 			.AddSingleton<IReplenishmentRepository, ReplenishmentRepository>()
 			.AddBatchService<IParcelDispatchService, ParcelDispatchService>()
 			.AddBatchService<IParcelGenerationService, ParcelGenerationService>()
-			.AddBatchService<IRequestReplenishmentService, RequestReplenishmentService>()
-			.AddMvc()
-			.SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+			.AddBatchService<IRequestReplenishmentService, RequestReplenishmentService>();
 	}
 }
