@@ -2,11 +2,8 @@
 using LogisticsSuite.WebShop.Handlers;
 using LogisticsSuite.WebShop.Services;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using IHostingEnvironment = Microsoft.AspNetCore.Hosting.IHostingEnvironment;
 
 namespace LogisticsSuite.WebShop
 {
@@ -17,24 +14,17 @@ namespace LogisticsSuite.WebShop
 		public Startup(IConfiguration configuration) => this.configuration = configuration;
 
 		// This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-		public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+		public void Configure(IApplicationBuilder app)
 		{
-			if (env.IsDevelopment())
-			{
-				app.UseDeveloperExceptionPage();
-			}
-
 			app.UseMessaging()
 				.Register<DelayChangeRequestedMessage, DelayChangeRequestedMessageHandler>();
 			app.UseBatchServices();
-			app.UseMvc();
 		}
 
 		// This method gets called by the runtime. Use this method to add services to the container.
 		public void ConfigureServices(IServiceCollection services) => services
 			.AddInfrastructure(configuration)
 			.AddBatchService<IWebOrderGenerationService, WebOrderGenerationService>()
-			.AddTransient<DelayChangeRequestedMessageHandler>()
-			.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+			.AddTransient<DelayChangeRequestedMessageHandler>();
 	}
 }
